change read in AudioSource to provide double

+++ AudioSource.hpp
@@ -4,7 +4,7 @@ class AudioSource {
   public:
     AudioSource() { }
     virtual ~AudioSource() { }
-    virtual void read(float *samples, size_t num) = 0;
+    virtual void read(double *samples, size_t num) = 0;
     virtual unsigned int getSampleRate() = 0;
 };
